1. Add a function call that updates the user's verification status when the user exits; 2. Move the identical code from the "Signup" and "Login" components to the "helper.js" file in a separate function; 3. Change a little and move the identical code from the "Signup" and "Login" components to the "hooks.js" file into a separate hook.
